module5-task1: * move redux types
