+++ src/common/enums/enums.ts
@@ -1,2 +1,3 @@
 export * from './app/app';
 export * from './game/game';
+export * from './redux/redux';
