+++ src/common/enums/redux/action-types/action-types.ts
@@ -1,0 +1,1 @@
+export * from './game';
