+++ src/common/enums/redux/action-types/game.ts
@@ -1,0 +1,7 @@
+enum GameActionType {
+  INCREMENT_STEP = `INCREMENT_STEP`,
+  RESET_GAME = `RESET_GAME`,
+  INCREMENT_MISTAKES = `INCREMENT_MISTAKES`,
+}
+
+export { GameActionType };
