+++ src/common/types/redux/action-cbs/action-cbs.ts
@@ -1,0 +1,1 @@
+export * from './game';
