+++ src/common/types/redux/action-cbs/game.ts
@@ -1,0 +1,17 @@
+import {
+  IncrementStepAction,
+  IncrementMistakeAction,
+  ResetGameAction,
+} from '~/common/types/redux/actions/game';
+
+type IncrementStepActionCb = (
+  incrementStepCount: number
+) => IncrementStepAction;
+
+type IncrementMistakeActionCb = (
+  incrementMistakeCount: number
+) => IncrementMistakeAction;
+
+type ResetGameActionCb = () => ResetGameAction;
+
+export { IncrementStepActionCb, IncrementMistakeActionCb, ResetGameActionCb };
