+++ src/common/types/redux/redux.ts
@@ -1,0 +1,2 @@
+export * from './actions/actions';
+export * from './action-cbs/action-cbs';
