+++ src/common/types/types.ts
@@ -3,3 +3,4 @@ export * from './game/game';
 export * from './callback/callback';
 export * from './event/event';
 export * from './hocs/hocs';
+export * from './redux/redux';
