+++ src/store/actions/actions.ts
@@ -1,0 +1,1 @@
+export * from './game/game';
