+++ src/store/reducer.root.ts
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import { gameReducer } from '~/store/reducers/game/game.reducer';
+import { gameReducer } from '~/store/reducers/game/game';
 
 const rootReducer = combineReducers({
   game: gameReducer,
